requirements.txt
flask_wtf
flask
email_validator
flask_session
flask_sqlalchemy
psycopg2
numpy
folium
pyepicollect
geopy
opencv-python
pandas
matplotlib
password_strength